+++ config/set/downgrade-php71.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-use Rector\DowngradePhp71\Rector\FunctionLike\DowngradeParamNullableTypeDeclarationRector;
-use Rector\DowngradePhp71\Rector\FunctionLike\DowngradeReturnNullableTypeDeclarationRector;
-use Rector\DowngradePhp71\Rector\FunctionLike\DowngradeReturnVoidTypeDeclarationRector;
+use Rector\DowngradePhp71\Rector\FunctionLike\DowngradeNullableTypeParamDeclarationRector;
+use Rector\DowngradePhp71\Rector\FunctionLike\DowngradeNullableTypeReturnDeclarationRector;
+use Rector\DowngradePhp71\Rector\FunctionLike\DowngradeVoidTypeReturnDeclarationRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
-    $services->set(DowngradeParamNullableTypeDeclarationRector::class);
-    $services->set(DowngradeReturnNullableTypeDeclarationRector::class);
-    $services->set(DowngradeReturnVoidTypeDeclarationRector::class);
+    $services->set(DowngradeNullableTypeParamDeclarationRector::class);
+    $services->set(DowngradeNullableTypeReturnDeclarationRector::class);
+    $services->set(DowngradeVoidTypeReturnDeclarationRector::class);
 };
